Add Airporter Demo

+++ modules/tb_functions.py
@@ -6,10 +6,11 @@ import time
 import os
 import glob
 
-from .utils import Config
+from .utils import Config, setup_logging
 from . import cc_functions
 
 config = Config()
+logger = setup_logging()
 
 def connection_create_kafka(
     kafka_bootstrap_servers,
@@ -45,61 +46,61 @@ def connection_create_kafka(
         data=json.dumps(connection_params)
     )
     resp.raise_for_status()
-    print(f"Created Kafka Connector named {kafka_connection_name} with id {resp.json()['id']}")
+    logger.info(f"Created Kafka Connector named {kafka_connection_name} with id {resp.json()['id']}")
 
 def connection_list():
-    print("Listing Tinybird Connections...")
+    logger.info("Listing Tinybird Connections...")
     resp = requests.get(
         config.TB_BASE_URL + "connectors",
         headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
     )
     resp.raise_for_status()
-    print(f"Found {len(resp.json()['connectors'])} Tinybird Connections.")
+    logger.info(f"Found {len(resp.json()['connectors'])} Tinybird Connections.")
     return resp.json()['connectors']
 
 def connection_delete(connection_id):
-    print(f"Deleting Kafka Connector with id {connection_id}")
+    logger.info(f"Deleting Kafka Connector with id {connection_id}")
     resp = requests.delete(
         config.TB_BASE_URL + "connectors/" + connection_id,
         headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
     )
     resp.raise_for_status()
-    print(f"Deleted Kafka Connector with id {connection_id}")
+    logger.info(f"Deleted Kafka Connector with id {connection_id}")
 
 def connectors_get(name, enforce_unique=True):
     # Note that connection names are not enforced unique by the service
-    print(f"Getting Tinybird Connectors with name {name}...")
+    logger.info(f"Getting Tinybird Connectors with name {name}...")
     connectors = connection_list()
     out = [connector for connector in connectors if connector['name'] == name]
     if enforce_unique:
         if len(out) > 1:
             raise Exception(f"Found {len(out)} Tinybird Connectors with name {name}, expected 1.")
     if out:
-        print(f"Found {len(out)} Tinybird Connectors with name {name}.")
+        logger.info(f"Found {len(out)} Tinybird Connectors with name {name}.")
         return out
-    print(f"Tinybird Connector(s) with name {name} not found.")
+    logger.info(f"Tinybird Connector(s) with name {name} not found.")
     return []
 
 def connection_test():
     connectors = connectors_get(name=config.TINYBIRD_CONFLUENT_CONNECTION_NAME)
     if not connectors:
-        print(f"Tinybird Confluent Connection {config.TINYBIRD_CONFLUENT_CONNECTION_NAME} not found.")
+        logger.info(f"Tinybird Confluent Connection {config.TINYBIRD_CONFLUENT_CONNECTION_NAME} not found.")
         return 0
     resp = requests.get(
         config.TB_BASE_URL + f"connectors/{connectors[0]['id']}/preview?preview_activity=false",
         headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
         )
     if not resp.ok:
-        print(f"Tinybird Confluent Connection {config.TINYBIRD_CONFLUENT_CONNECTION_NAME} not working: {resp.text}")
+        logger.info(f"Tinybird Confluent Connection {config.TINYBIRD_CONFLUENT_CONNECTION_NAME} not working: {resp.text}")
         return -1
     
-    print(f"Tinybird Confluent Connection has {len(resp.json()['preview'])} Topics.")
+    logger.info(f"Tinybird Confluent Connection has {len(resp.json()['preview'])} Topics.")
     return 1
 
 def ensure_kafka_connection():
     status = connection_test()
     if status == 0:
-        print("Tinybird Confluent Connection not found. Creating...")
+        logger.info("Tinybird Confluent Connection not found. Creating...")
         connection_create_kafka(
             kafka_bootstrap_servers=config.CONFLUENT_BOOTSTRAP_SERVERS,
             kafka_key=config.CONFLUENT_UNAME,
@@ -107,7 +108,7 @@ def ensure_kafka_connection():
             kafka_connection_name=config.TINYBIRD_CONFLUENT_CONNECTION_NAME,
         )
     elif status == -1:
-        print("Tinybird Confluent Connection not working. Deleting and creating...")
+        logger.info("Tinybird Confluent Connection not working. Deleting and creating...")
         connector = connectors_get(name=config.TINYBIRD_CONFLUENT_CONNECTION_NAME)
         connection_delete(connector[0]['id'])
         time.sleep(5)
@@ -121,16 +122,16 @@ def ensure_kafka_connection():
         if not connection_test():
             raise Exception("Tinybird Confluent Connection not working after recreation.")
     else:
-        print(f"Tinybird Confluent Connection {config.TINYBIRD_CONFLUENT_CONNECTION_NAME} found.")
+        logger.info(f"Tinybird Confluent Connection {config.TINYBIRD_CONFLUENT_CONNECTION_NAME} found.")
 
 def datasources_list():
-    print(f"Listing Tinybird Datasources...")
+    logger.info(f"Listing Tinybird Datasources...")
     resp = requests.get(
         config.TB_BASE_URL + "datasources",
         headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
     )
     resp.raise_for_status()
-    print(f"Found {len(resp.json()['datasources'])} Tinybird Datasources.")
+    logger.info(f"Found {len(resp.json()['datasources'])} Tinybird Datasources.")
     return resp.json()['datasources']
 
 def datasources_truncate(names, include_quarantine=True):
@@ -142,31 +143,31 @@ def datasources_truncate(names, include_quarantine=True):
             ds_names = [name]
     for ds_name in ds_names:
         if ds_name not in [x['name'] for x in ds_list]:
-            print(f"Tinybird Datasource {ds_name} not found.")
+            logger.info(f"Tinybird Datasource {ds_name} not found.")
         else:
-            print(f"Truncating Tinybird Datasource {ds_name}")
+            logger.info(f"Truncating Tinybird Datasource {ds_name}")
             resp = requests.post(
                 config.TB_BASE_URL + f"datasources/{ds_name}/truncate",
                 headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
             )
             resp.raise_for_status()
-            print(f"Truncated Tinybird Datasource {ds_name}")
+            logger.info(f"Truncated Tinybird Datasource {ds_name}")
 
 def datasources_delete(names):
     ds_list = datasources_list()
     for name in names:
         if name not in [x['name'] for x in ds_list]:
-            print(f"Tinybird Datasource {name} not found.")
+            logger.info(f"Tinybird Datasource {name} not found.")
         else:
-            print(f"Deleting Tinybird Datasource {name}")
+            logger.info(f"Deleting Tinybird Datasource {name}")
             resp = requests.delete(
                 config.TB_BASE_URL + f"datasources/{name}",
                 headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
             )
             resp.raise_for_status()
-            print(f"Deleted Tinybird Datasource {name}")
+            logger.info(f"Deleted Tinybird Datasource {name}")
 
-def update_datasource_info(files, kafka_topic):
+def update_datasource_info(files):
     for filepath in files:
         if 'raw.datasource' in filepath:
             # Replace the Kafka Group ID in the definition files
@@ -174,18 +175,11 @@ def update_datasource_info(files, kafka_topic):
             with open(filepath, 'r') as file:
                 file_content = file.read()
 
-            print(f"Generating new Kafka Group ID in {filename}...")
-            # Replace the KAFKA_GROUP_ID line
-            start_index = file_content.find("KAFKA_GROUP_ID '") + len("KAFKA_GROUP_ID '")
-            end_index = file_content.find("'", start_index)
-            old_group_id = file_content[start_index:end_index]
-            # Generate new group_id with current unix time
-            new_group_id = old_group_id[:old_group_id.rfind("_")+1] + str(int(time.time()))
-            file_content = file_content.replace(old_group_id, new_group_id)
-            print(f"Generated new Kafka Group ID: {new_group_id}")
-
             # Ensure KAFKA_TOPIC is correct for database and table name
-            print(f"Updating Kafka Topic to {kafka_topic}...")
+            # Expecting file to be named in the format <table_name>_raw.datasource
+            table_name = filename.split('_')[0]
+            kafka_topic = config.KAFKA_CDC_TOPICS[table_name]
+            logger.info(f"Updating Kafka Topic to {kafka_topic}...")
             topic_start_index = file_content.find("KAFKA_TOPIC '") + len("KAFKA_TOPIC '")
             topic_end_index = file_content.find("'", topic_start_index)
             old_topic = file_content[topic_start_index:topic_end_index]
@@ -193,26 +187,36 @@ def update_datasource_info(files, kafka_topic):
             if len(topics) != 1:
                 raise Exception(f"Found {len(topics)} topics matching {kafka_topic}, expected 1.")
             else:
-                print(f"Found Kafka Topic {topics[0]} matching {kafka_topic}")
+                logger.info(f"Found Kafka Topic {topics[0]} matching {kafka_topic}")
                 new_topic = topics[0]
             file_content = file_content.replace(old_topic, new_topic)
-            print(f"Updated Kafka Topic to {new_topic}")
+            logger.info(f"Updated Kafka Topic to {new_topic}")
+
+            logger.info(f"Generating new Kafka Group ID in {filename}...")
+            # Replace the KAFKA_GROUP_ID line
+            start_index = file_content.find("KAFKA_GROUP_ID '") + len("KAFKA_GROUP_ID '")
+            end_index = file_content.find("'", start_index)
+            old_group_id = file_content[start_index:end_index]
+            # Generate new group_id with current unix time
+            new_group_id = new_topic + '_' + str(int(time.time()))
+            file_content = file_content.replace(old_group_id, new_group_id)
+            logger.info(f"Generated new Kafka Group ID: {new_group_id}")
             
             # Fix KAFKA_AUTO_OFFSET_RESET as well
-            print(f"Updating Kafka Auto Offset Reset to {config.CONFLUENT_OFFSET_RESET}...")
+            logger.info(f"Updating Kafka Auto Offset Reset to {config.CONFLUENT_OFFSET_RESET}...")
             offset_start_index = file_content.find("KAFKA_AUTO_OFFSET_RESET '") + len("KAFKA_AUTO_OFFSET_RESET '")
             offset_end_index = file_content.find("'", offset_start_index)
             old_offset = file_content[offset_start_index:offset_end_index]
             file_content = file_content.replace(old_offset, config.CONFLUENT_OFFSET_RESET)
-            print(f"Updated Kafka Auto Offset Reset to {config.CONFLUENT_OFFSET_RESET}")
+            logger.info(f"Updated Kafka Auto Offset Reset to {config.CONFLUENT_OFFSET_RESET}")
 
             # Fix KAFKA_CONNECTION_NAME
-            print(f"Updating Kafka Connection Name to {config.TINYBIRD_CONFLUENT_CONNECTION_NAME}...")
+            logger.info(f"Updating Kafka Connection Name to {config.TINYBIRD_CONFLUENT_CONNECTION_NAME}...")
             connection_start_index = file_content.find("KAFKA_CONNECTION_NAME '") + len("KAFKA_CONNECTION_NAME '")
             connection_end_index = file_content.find("'", connection_start_index)
             old_connection = file_content[connection_start_index:connection_end_index]
             file_content = file_content.replace(old_connection, config.TINYBIRD_CONFLUENT_CONNECTION_NAME)
-            print(f"Updated Kafka Connection Name to {config.TINYBIRD_CONFLUENT_CONNECTION_NAME}")
+            logger.info(f"Updated Kafka Connection Name to {config.TINYBIRD_CONFLUENT_CONNECTION_NAME}")
 
             # Write the modified content back to the file
             with open(filepath, 'w') as file:
@@ -221,54 +225,45 @@ def update_datasource_info(files, kafka_topic):
 def upload_def_file(filepath):
     url = config.TB_BASE_URL + "datafiles"
     filename = os.path.basename(filepath)
-    print(f"Reading in file {filepath}...")
+    logger.info(f"Reading in file {filepath}...")
     file_data = {filename: open(filepath, 'rb')}
     headers = {
         "Authorization": f"Bearer {config.TINYBIRD_API_KEY}",
     }
-    print(f"Uploading {filepath} to Tinybird...")
+    logger.info(f"Uploading {filepath} to Tinybird...")
     resp = requests.post(
         url + f"?filenames={filename}",
         headers=headers,
         files=file_data,
     )
     if resp.ok:
-        print(f"Successfully uploaded {filename} to Tinybird.")
+        logger.info(f"Successfully uploaded {filename} to Tinybird.")
     elif resp.status_code == 400:
-        print(f"Tinybird returned an Error: {resp.text}")
+        logger.info(f"Tinybird returned an Error: {resp.text}")
+        resp.raise_for_status()
     else:
         resp.raise_for_status()
-
-def get_def_files_for_db(source_db):
-    print(f"Getting Tinybird definitions for {source_db}...")
-    files_to_get = [
-        filepath 
-        for directory in ["./datasources", "./pipes"]
-        for filepath in glob.glob(os.path.join(directory, f"{source_db.lower()}_*"))
-    ]
-    print(f"Got {len(files_to_get)} Tinybird definitions for {source_db}.")
-    return files_to_get
 
 def upload_def_for_db(files):
     # We deliberately use the Tinybird definition files here to replicate what a user would typically do in the CLI.
     # It avoids the hassle of converting the schemas and definitions into plain python objects to POST with requests.
     # Upload files
     _ = [upload_def_file(filepath) for filepath in files]
-    print(f"Uploaded {len(files)} Tinybird definitions.")
+    logger.info(f"Uploaded {len(files)} Tinybird definitions.")
 
 def pipes_list():
-    print(f"Listing Tinybird Pipes...")
+    logger.info(f"Listing Tinybird Pipes...")
     resp = requests.get(
         config.TB_BASE_URL + "pipes",
         headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
     )
     resp.raise_for_status()
-    print(f"Found {len(resp.json()['pipes'])} Tinybird Pipes.")
+    logger.info(f"Found {len(resp.json()['pipes'])} Tinybird Pipes.")
     return resp.json()['pipes']
 
 def endpoint_fetch(endpoint_name):
     # e.g. https://api.tinybird.co/v0/pipes/pg_users_api_rmt.json
-    print(f"Fetching Tinybird Endpoint {endpoint_name}...")
+    logger.info(f"Fetching Tinybird Endpoint {endpoint_name}...")
     resp = requests.get(
         config.TB_BASE_URL + f"pipes/{endpoint_name}",
         headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
@@ -280,54 +275,47 @@ def pipes_delete(names):
     pipes_listing = pipes_list()
     for name in names:
         if name not in [x['name'] for x in pipes_listing]:
-            print(f"Tinybird Pipe {name} not found.")
+            logger.info(f"Tinybird Pipe {name} not found.")
         else:
-            print(f"Deleting Tinybird Pipe {name}")
+            logger.info(f"Deleting Tinybird Pipe {name}")
             resp = requests.delete(
                 config.TB_BASE_URL + f"pipes/{name}",
                 headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
             )
             resp.raise_for_status()
-            print(f"Deleted Tinybird Pipe {name}")
+            logger.info(f"Deleted Tinybird Pipe {name}")
 
-def clean_workspace(source_db, include_connector=False):
-    print(f"Cleaning Tinybird workspace for {source_db}...")
-    files = get_def_files_for_db(source_db)
-    pipes_to_delete = [file for file in files if os.path.dirname(file) == './pipes']
-    datasources_to_delete = [file for file in files if os.path.dirname(file) == './datasources']
+def clean_workspace(files, include_connector=False):
+    pipes_to_delete = [file for file in files if '/pipes' in os.path.dirname(file)]
+    datasources_to_delete = [file for file in files if '/datasources' in os.path.dirname(file)]
     # Do Pipes first
-    current_pipe_names = [x['name'] for x in pipes_list()]
     pipenames_to_delete = [os.path.basename(file).split('.')[0] for file in pipes_to_delete]
-    pipes_delete([x for x in pipenames_to_delete if x in current_pipe_names])
+    logger.info(f"Requesting Delete of Pipes: {pipenames_to_delete}")
+    pipes_delete(pipenames_to_delete)
     # Then do Datasources
-    current_datasource_names = [x['name'] for x in datasources_list()]
     datasourcenames_to_delete = [os.path.basename(file).split('.')[0] for file in datasources_to_delete]
-    datasources_delete([x for x in datasourcenames_to_delete if x in current_datasource_names])
-    # Then do Dataspace Tokens
-    current_tokens = tokens_list()
-    current_token_names = [x['name'] for x in current_tokens]
-    kafka_token_patten = "_".join([
-        config.TINYBIRD_CONFLUENT_CONNECTION_NAME,
-        source_db.lower(),
-        config.USERS_TABLE_NAME
-        ])
-    tokens_to_remove = [
-        token_name for token_name in current_token_names if token_name.startswith(kafka_token_patten)
-        ]
-    # Then do Pipe Tokens
-    tokens_to_remove += [
-        x for x in get_token_names_from_pipes(pipes_to_delete) if x in current_token_names
-        ]
-    print(f"Found {len(tokens_to_remove)} Tinybird Tokens to remove for {source_db}.")
+    logger.info(f"Requesting Delete of Datasources: {datasourcenames_to_delete}")
+    datasources_delete(datasourcenames_to_delete)
+    # Then do Workspace Tokens
+    tokens_to_remove = []
+    # First Datasources
+    for datasource_name in datasourcenames_to_delete:
+        tokens_to_remove.append("_".join([
+            config.TINYBIRD_CONFLUENT_CONNECTION_NAME,
+            datasource_name
+            ]))
+    # Then add Pipe Tokens
+    tokens_to_remove += get_token_names_from_pipes(pipes_to_delete)
+    logger.info(f"Requesting Delete of Tokens: {tokens_to_remove}")
     # Do token removal
-    _ = [tokens_delete(token_name) for token_name in set(tokens_to_remove)]
+    tokens_delete(set(tokens_to_remove))
     if include_connector:
         # Do Connectors
         connector = connectors_get(name=config.TINYBIRD_CONFLUENT_CONNECTION_NAME)
         if connector:
             connection_delete(connector[0]['id'])
         else:
-            print(f"Tinybird Confluent Connection not found for name: {config.TINYBIRD_CONFLUENT_CONNECTION_NAME}")
+            logger.info(f"Tinybird Confluent Connection not found for name: {config.TINYBIRD_CONFLUENT_CONNECTION_NAME}")
 
 def get_token_names_from_pipes(pipes):
     token_names = []
@@ -341,21 +329,26 @@ def get_token_names_from_pipes(pipes):
     return token_names
 
 def tokens_list():
-    print(f"Listing Tinybird Tokens...")
+    logger.info(f"Listing Tinybird Tokens...")
     resp = requests.get(
         config.TB_BASE_URL + "tokens",
         headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
     )
     resp.raise_for_status()
-    print(f"Found {len(resp.json()['tokens'])} Tinybird Tokens.")
+    logger.info(f"Found {len(resp.json()['tokens'])} Tinybird Tokens.")
     return resp.json()['tokens']
 
-def tokens_delete(name):
-    print(f"Deleting Tinybird Token {name}")
-    resp = requests.delete(
-        config.TB_BASE_URL + f"tokens/{name}",
-        headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
-    )
-    resp.raise_for_status()
-    print(f"Deleted Tinybird Token {name}")
+def tokens_delete(names):
+    token_list = tokens_list()
+    for name in names:
+        if name not in [x['name'] for x in token_list]:
+            logger.info(f"Tinybird Token {name} not found.")
+        else:
+            logger.info(f"Deleting Tinybird Token {name}")
+            resp = requests.delete(
+                config.TB_BASE_URL + f"tokens/{name}",
+                headers={'Authorization': f'Bearer {config.TINYBIRD_API_KEY}'}
+            )
+            resp.raise_for_status()
+            logger.info(f"Deleted Tinybird Token {name}")
